Review notes addressed
