+++ src/core/i18n/en/translation.en.json
@@ -791,7 +791,6 @@
     }
   },
   "components": {
-    "myDashboard": "My Dashboard",
     "search": {
       "scope": {
         "platform": "Entire platform",
@@ -1830,6 +1829,7 @@
       "innovation-flow-info": "Click to display the Innovation Flow"
     },
     "home": {
+      "title": "My Dashboard",
       "sections": {
         "welcome": {
           "welcome-back": "Welcome back {{username}}!",
